+++ src/main/java/com/example/community/model/Topic.java
@@ -5,13 +5,13 @@ import lombok.Data;
 // topic information from database
 @Data
 public class Topic {
-    private Integer id;
+    private Long id;
     private String title;
     private String description;
     private String tag;
     private Long gmtCreate;
     private Long gmtModified;
-    private Integer author;
+    private Long author;
     private Integer viewCount;
     private Integer commentCount;
     private Integer likeCount;
